requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor-5==0.2.18
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-recaptcha==3.0.0
django-widget-tweaks==1.5.0
environs==14.2.0
gunicorn==23.0.0
marshmallow==4.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
